chore: better naming and re-org of files
